healthserver/nutrition-service/requirements.txt
# # # # absl-py==2.3.1
# # # # argon2-cffi==25.1.0
# # # # argon2-cffi-bindings==25.1.0
# # # # astunparse==1.6.3
# # # # blinker==1.9.0
# # # # certifi==2025.10.5
# # # # cffi==2.0.0
# # # # charset-normalizer==3.4.4
# # # # click==8.3.0
# # # # colorama==0.4.6
# # # # dnspython==2.8.0
# # # # Flask==3.1.2
# # # # flask-cors==6.0.1
# # # # flatbuffers==25.9.23
# # # # gast==0.6.0
# # # # google-pasta==0.2.0
# # # # grpcio==1.76.0
# # # # h5py==3.15.1
# # # # idna==3.11
# # # # itsdangerous==2.2.0
# # # # Jinja2==3.1.6
# # # # keras==3.12.0
# # # # libclang==18.1.1
# # # # Markdown==3.9
# # # # markdown-it-py==4.0.0
# # # # MarkupSafe==3.0.3
# # # # mdurl==0.1.2
# # # # minio==7.2.18
# # # # ml_dtypes==0.5.3
# # # # namex==0.1.0
# # # # numpy==2.2.6
# # # # opencv-python==4.12.0.88
# # # # opt_einsum==3.4.0
# # # # optree==0.17.0
# # # # packaging==25.0
# # # # pillow==12.0.0
# # # # protobuf==6.33.0
# # # # pycparser==2.23
# # # # pycryptodome==3.23.0
# # # # Pygments==2.19.2
# # # # PyJWT==2.10.1
# # # # pymongo==4.15.3
# # # # python-dotenv==1.2.1
# # # # requests==2.32.5
# # # # rich==14.2.0
# # # # six==1.17.0
# # # # tensorboard==2.20.0
# # # # tensorboard-data-server==0.7.2
# # # # tensorflow==2.20.0
# # # # termcolor==3.2.0
# # # # typing_extensions==4.15.0
# # # # urllib3==2.5.0
# # # # Werkzeug==3.1.3
# # # # wrapt==2.0.0
# # # # ============================================
# # # # DÉPENDANCES EXISTANTES (Flask, etc.)
# # # # ============================================
# # # absl-py==2.3.1
# # # argon2-cffi==25.1.0
# # # argon2-cffi-bindings==25.1.0
# # # astunparse==1.6.3
# # # blinker==1.9.0
# # # certifi==2025.10.5
# # # cffi==2.0.0
# # # charset-normalizer==3.4.4
# # # click==8.3.0
# # # colorama==0.4.6
# # # dnspython==2.8.0
# # # Flask==3.1.2
# # # flask-cors==6.0.1
# # # flatbuffers==25.9.23
# # # gast==0.6.0
# # # google-pasta==0.2.0
# # # grpcio==1.76.0
# # # idna==3.11
# # # itsdangerous==2.2.0
# # # Jinja2==3.1.6
# # # libclang==18.1.1
# # # Markdown==3.9
# # # markdown-it-py==4.0.0
# # # MarkupSafe==3.0.3
# # # mdurl==0.1.2
# # # minio==7.2.18
# # # namex==0.1.0
# # # opencv-python==4.12.0.88
# # # packaging==25.0
# # # pillow==12.0.0
# # # protobuf==6.33.0
# # # pycparser==2.23
# # # pycryptodome==3.23.0
# # # Pygments==2.19.2
# # # PyJWT==2.10.1
# # # pymongo==4.15.3
# # # python-dotenv==1.2.1
# # # requests==2.32.5
# # # rich==14.2.0
# # # six==1.17.0
# # # termcolor==3.2.0
# # # urllib3==2.5.0
# # # Werkzeug==3.1.3

# # # # ============================================
# # # # ⭐ NOUVELLES DÉPENDANCES POUR LE MODÈLE AI
# # # # ============================================

# # # # TensorFlow (pour charger et utiliser le modèle)
# # # tensorflow==2.15.0
# # # # OU si vous n'avez pas de GPU:
# # # # tensorflow-cpu==2.15.0

# # # # Pour la gestion des arrays numpy (déjà présent mais version spécifique)
# # # numpy==1.24.3

# # # # h5py pour charger les modèles .h5
# # # h5py==3.15.1

# # # # Keras (inclus dans TensorFlow mais explicite)
# # # keras==3.12.0

# # # # ML/Data Science utilities
# # # ml_dtypes==0.5.3
# # # opt_einsum==3.4.0
# # # optree==0.17.0

# # # # TensorBoard (optionnel - pour monitoring)
# # # tensorboard==2.20.0
# # # tensorboard-data-server==0.7.2

# # # # Autres dépendances ML
# # # wrapt==2.0.0
# # # typing_extensions==4.15.0



# # # ============================================
# # # 🌐 DEPENDANCES FLASK & UTILITAIRES
# # # ============================================
# # Flask==3.1.2
# # flask-cors==6.0.1
# # requests==2.32.5
# # pymongo==4.15.3
# # python-dotenv==1.2.1
# # Werkzeug==3.1.3
# # itsdangerous==2.2.0
# # Jinja2==3.1.6
# # MarkupSafe==3.0.3
# # colorama==0.4.6
# # click==8.1.7
# # rich==14.2.0

# # # ============================================
# # # 🗄️ MINIO & CRYPTOGRAPHIE
# # # ============================================
# # minio==7.2.18
# # pycryptodome==3.23.0

# # # ============================================
# # # 🤖 MACHINE LEARNING & INTELLIGENCE ARTIFICIELLE
# # # ============================================
# # # tensorflow==2.15.0
# # # tensorflow-intel==2.16.1
# # tensorflow==2.16.1

# # numpy==1.24.3
# # opencv-python==4.9.0.80

# # h5py==3.15.1
# # pillow==12.0.0
# # keras==3.12.0

# # # ============================================
# # # ⚙️ DEPENDANCES TECHNIQUES TENSORFLOW
# # # ============================================
# # absl-py==2.3.1
# # astunparse==1.6.3
# # flatbuffers==25.9.23
# # gast==0.6.0
# # google-pasta==0.2.0
# # grpcio==1.76.0
# # ml_dtypes==0.5.3
# # opt_einsum==3.4.0
# # optree==0.17.0
# # tensorboard==2.15.0
# # tensorboard-data-server==0.7.2
# # termcolor==3.2.0
# # typing_extensions==4.15.0
# # wrapt==2.0.0

# # # ============================================
# # # 📦 AUTRES LIBRAIRIES UTILES
# # # ============================================
# # # protobuf==6.33.0
# # protobuf==3.20.3

# # packaging==25.0
# # idna==3.11
# # charset-normalizer==3.4.4
# # urllib3==2.5.0
# # certifi==2025.10.5
# # six==1.17.0
# # dnspython==2.8.0



# # ============================================
# # 🌐 DEPENDANCES FLASK & UTILITAIRES
# # ============================================
# Flask==3.1.2
# flask-cors==6.0.1
# requests==2.32.5
# pymongo==4.15.3
# python-dotenv==1.2.1
# Werkzeug==3.1.3
# itsdangerous==2.2.0
# Jinja2==3.1.6
# MarkupSafe==3.0.3
# colorama==0.4.6
# click==8.1.7
# rich==14.2.0

# # ============================================
# # 🗄️ MINIO & CRYPTOGRAPHIE
# # ============================================
# minio==7.2.18
# pycryptodome==3.23.0

# # ============================================
# # 🤖 MACHINE LEARNING & INTELLIGENCE ARTIFICIELLE
# # ============================================
# tensorflow==2.16.1
# numpy==1.24.3
# opencv-python==4.9.0.80
# h5py==3.15.1
# pillow==12.0.0
# keras==3.12.0

# # ============================================
# # ⚙️ DEPENDANCES TECHNIQUES TENSORFLOW
# # ============================================
# absl-py==2.3.1
# astunparse==1.6.3
# flatbuffers==25.9.23
# gast==0.6.0
# google-pasta==0.2.0
# grpcio==1.76.0
# ml_dtypes==0.5.3
# opt_einsum==3.4.0
# optree==0.17.0
# tensorboard==2.16.1
# tensorboard-data-server==0.8.1
# termcolor==3.2.0
# typing_extensions==4.15.0
# wrapt==2.0.0
# protobuf==3.20.3

# # ============================================
# # 📦 AUTRES LIBRAIRIES UTILES
# # ============================================
# packaging==25.0
# idna==3.11
# charset-normalizer==3.4.4
# urllib3==2.5.0
# certifi==2025.10.5
# six==1.17.0
# dnspython==2.8.0



# DEPENDANCES FLASK & UTILITAIRES

Flask==3.1.2
flask-cors==6.0.1
requests==2.32.5
pymongo==4.15.3
python-dotenv==1.2.1
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
colorama==0.4.6
click==8.1.7
rich==14.2.0
blinker>=1.9.0
markdown-it-py>=2.2.0
pygments>=2.13.0,<3.0.0

# MINIO & CRYPTOGRAPHIE

minio==7.2.18
pycryptodome==3.23.0
argon2-cffi

# MACHINE LEARNING & IA

tensorflow==2.16.1
numpy==1.24.3
opencv-python==4.9.0.80
h5py==3.15.1
pillow==12.0.0
keras==3.12.0

# DEPENDANCES TECHNIQUES TENSORFLOW

absl-py==2.3.1
astunparse==1.6.3
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0

opt_einsum==3.4.0
optree==0.17.0
tensorboard==2.16.1
tensorboard-data-server==0.7.2
termcolor==3.2.0
typing_extensions==4.15.0
wrapt==2.0.0
protobuf==3.20.3

# AUTRES LIBRAIRIES UTILES
packaging==25.0
idna==3.11
charset-normalizer==3.4.4
urllib3==2.5.0
certifi==2025.10.5
six==1.17.0
dnspython==2.8.0
